Update dependencies to Laravel 5.5, integrate Passport and do some more structure
Passport for API Authentication was integrated but is not in use yet.
Some models were added for testing purposes (but since they are related to the use of this system, they can be extended for the final product)
All dependencies are now updated to be used with PHP 7 and Laravel 5.5

Attention: this commit still has a bug in the migrations. The 2017_10_12_225441_foreign_keys_for_flight_arrival_and_departure.php migration does not complete since it complains about a missing field. I wasn't able to figure out why the field should be missing.